modified changed CJSON_GET_VALUE_INT to CJSON_GET_VALUE_DOUBLE

+++ ezlopi-cloud/ezlopi-cloud-coordinates/ezlopi_cloud_coordinates.h
@@ -4,10 +4,10 @@
 
 #include "cJSON.h"
 
-void hub_coordinates_set(cJSON *cjson_request, cJSON *cjson_response);
+void hub_coordinates_set(cJSON* cjson_request, cJSON* cjson_response);
 void hub_coordinates_get(cJSON* cj_request, cJSON* cj_response);
-float ezlopi_cloud_get_longitude();
-float ezlopi_cloud_get_latitude();
+double ezlopi_cloud_get_longitude();
+double ezlopi_cloud_get_latitude();
 
 #endif // _EZLOPI_CLOUD_COORDINSTES_H_
